Moves selectors folder into public and deletes store folder
