+++ x-pack/plugins/apm/public/components/shared/charts/CustomPlot/test/CustomPlot.test.js
@@ -16,7 +16,7 @@ import InteractivePlot from '../InteractivePlot';
 import {
   getResponseTimeSeries,
   getEmptySerie
-} from '../../../../../store/selectors/chartSelectors';
+} from '../../../../../selectors/chartSelectors';
 
 function getXValueByIndex(index) {
   return responseWithData.responseTimes.avg[index].x;
