+++ x-pack/plugins/apm/public/components/shared/charts/TransactionCharts/index.tsx
@@ -19,7 +19,7 @@ import React, { Component } from 'react';
 import { isEmpty } from 'lodash';
 import styled from 'styled-components';
 import { Coordinate } from '../../../../../typings/timeseries';
-import { ITransactionChartData } from '../../../../store/selectors/chartSelectors';
+import { ITransactionChartData } from '../../../../selectors/chartSelectors';
 import { IUrlParams } from '../../../../context/UrlParamsContext/types';
 import { asInteger, asMillis, tpmUnit } from '../../../../utils/formatters';
 import { LicenseContext } from '../../../app/Main/LicenseCheck';
