+++ x-pack/plugins/apm/public/selectors/__tests__/mockData/anomalyData.ts
